python-agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
httpx>=0.27.0

# LLM — let langchain resolve its own dependency tree
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# Tracing
langsmith>=0.1.100
langfuse>=2.44.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Charts
matplotlib>=3.9.0
numpy>=1.26.0
Pillow>=10.4.0

# Vector store
chromadb>=0.5.0
